task 6up

+++ 0x00-hello_world/6-size.c
@@ -3,7 +3,7 @@
 /**
  * main - Prints the size of different variable types
  *
- * Return: Always 0 (good)
+ * Return: Always 0 (Successful)
  */
 int main(void)
 {
@@ -13,10 +13,10 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of a char: %d byte(s)\n",(unsigned long)sizeof(a));
-        printf("Size of an int: %d byte(s)\n",(unsigned long)sizeof(b));
-	printf("Size of a long int: % byte(s)\n",(unsigned long)sizeof(c));
-	printf("Size of a long long int: % byte(s)\n",(unsigned long)sizeof(d));
-	printf("Size of a float: %d byte(s)\n",(unsigned long)sizeof(e);
+	printf("Size of a char: %lu byte(s)\n", sizeof(a));
+        printf("Size of an int: %lu byte(s)\n", sizeof(b));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(c));
+	printf("Size of a long long int: %lu byte(s)\n", sizeof(d));
+	printf("Size of a float: %lu byte(s)\n", sizeof(e));
         return (0);
-	}
+}
